bilibili-goods/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
streamlit>=1.49.0
pandas>=2.3.0
openpyxl>=3.1.5
requests>=2.32.0
httpx>=0.28.0
bilibili-api-python>=17.3.0
playwright>=1.55.0
python-dotenv>=1.1.0
